# packages/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for Sistema de Controle de Visitantes
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /visits:
    get:
      operationId: listVisits
      tags: [visits]
      summary: List visits with filters
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: sectorId
          schema:
            type: integer
        - in: query
          name: status
          schema:
            type: string
            enum: [ongoing, finished, cancelled]
        - in: query
          name: dateFrom
          schema:
            type: string
            format: date
        - in: query
          name: dateTo
          schema:
            type: string
            format: date
        - in: query
          name: userId
          schema:
            type: integer
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of visits
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VisitListResponse"
    post:
      operationId: createVisit
      tags: [visits]
      summary: Register a visit entry (creates visitor if not exists)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VisitInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Visit"
  /visits/{id}:
    get:
      operationId: getVisit
      tags: [visits]
      summary: Get visit by ID
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Visit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Visit"
    patch:
      operationId: updateVisit
      tags: [visits]
      summary: Update visit (admin only - corrections)
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VisitUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Visit"
  /visits/{id}/checkout:
    post:
      operationId: checkoutVisit
      tags: [visits]
      summary: Register visit exit
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Checked out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Visit"
  /visits/{id}/cancel:
    post:
      operationId: cancelVisit
      tags: [visits]
      summary: Cancel a visit (admin only)
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CancelVisitInput"
      responses:
        "200":
          description: Cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Visit"
  /visits/{id}/reprint:
    post:
      operationId: reprintLabel
      tags: [visits]
      summary: Log label reprint (admin only)
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Reprinted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ── Dashboard ─────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdParam:
      in: path
      name: id
      required: true
      schema:
        type: integer
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success

    # ── Auth ──────────────────────────────────────────────────────────────
...
    User:
      type: object
      required: [id, name, login, role, status, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        login:
          type: string
        role:
          type: string
          enum: [admin, receptionist]
        status:
          type: string
          enum: [active, inactive]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: ["string", "null"]
          format: date-time
...
    Sector:
      type: object
      required: [id, name, abbreviation, secretariat, status, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        abbreviation:
          type: string
        secretariat:
          type: string
        status:
          type: string
          enum: [active, inactive]
        createdAt:
          type: string
          format: date-time
...
    Visitor:
      type: object
      required: [id, name, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        cpf:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        company:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: ["string", "null"]
          format: date-time
...
    Visit:
      type: object
      required: [id, visitorId, sectorId, status, entryDate, entryTime, entryUserId, createdAt]
      properties:
        id:
          type: integer
        visitorId:
          type: integer
        visitor:
          $ref: "#/components/schemas/Visitor"
        sectorId:
          type: integer
        sector:
          $ref: "#/components/schemas/Sector"
        responsible:
          type: ["string", "null"]
        reason:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        status:
          type: string
          enum: [ongoing, finished, cancelled]
        entryDate:
          type: string
          format: date
        entryTime:
          type: string
        entryUserId:
          type: integer
        entryUser:
          $ref: "#/components/schemas/User"
        exitDate:
          type: ["string", "null"]
          format: date
        exitTime:
          type: ["string", "null"]
        exitUserId:
          type: ["integer", "null"]
        exitUser:
          $ref: "#/components/schemas/User"
        cancelReason:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    VisitInput:
      type: object
      required: [sectorId]
      properties:
        visitorId:
          type: ["integer", "null"]
        visitorName:
          type: string
        visitorCpf:
          type: string
        visitorPhone:
          type: string
        visitorCompany:
          type: string
        visitorCity:
          type: string
        updateVisitorData:
          type: boolean
          default: false
        sectorId:
          type: integer
        responsible:
          type: string
        reason:
          type: string
        notes:
          type: string
    VisitUpdate:
      type: object
      properties:
        sectorId:
          type: integer
        responsible:
          type: string
        reason:
          type: string
        notes:
          type: string
        exitDate:
          type: string
          format: date
        exitTime:
          type: string
    CancelVisitInput:
      type: object
      required: [reason]
      properties:
        reason:
          type: string
    VisitListResponse:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Visit"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    # ── Dashboard ─────────────────────────────────────────────────────────
...