# packages/api-zod/src/generated/types/visitUpdate.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification for Sistema de Controle de Visitantes
 * OpenAPI spec version: 0.1.0
 */

export interface VisitUpdate {
  sectorId?: number;
  responsible?: string;
  reason?: string;
  notes?: string;
  exitDate?: Date;
  exitTime?: string;
}